improvement: log action output to cli consistently
Change the ActionRouter to not only forward logs to Garden Cloud,
but also send logs to the CLI in a way that is consistent between
plugins.
