+++ core/src/plugins/container/build.ts
@@ -11,8 +11,6 @@ import { ContainerModule } from "./config"
 import { ConfigurationError } from "../../exceptions"
 import { GetBuildStatusParams } from "../../types/plugin/module/getBuildStatus"
 import { BuildModuleParams } from "../../types/plugin/module/build"
-import { LogLevel } from "../../logger/logger"
-import { renderOutputStream } from "../../util/util"
 import { PrimitiveMap } from "../../config/common"
 import split2 from "split2"
 
@@ -67,12 +65,9 @@ export async function buildContainerModule({ ctx, module, log }: BuildModulePara
 
   // Stream verbose log to a status line
   const outputStream = split2()
-  const statusLine = log.placeholder({ level: LogLevel.verbose })
-
   outputStream.on("error", () => {})
   outputStream.on("data", (line: Buffer) => {
     ctx.events.emit("log", { timestamp: new Date().getTime(), data: line })
-    statusLine.setState(renderOutputStream(line.toString(), "local-docker"))
   })
   const timeout = module.spec.build.timeout
   const res = await containerHelpers.dockerCli({
